refactor: useDiaryEmotionStore를 store 레이어로 이동 (#30)
features 간 교차 참조를 해소하기 위해 store/ 레이어로 이동
