+++ src/features/diary-emotion/ui/SituationStep.tsx
@@ -3,10 +3,10 @@
 import { useEffect } from "react";
 
 import { useEmotionTagsQuery } from "@/entities/emotion-tag";
+import { MOCK_SITUATIONS } from "@/mock";
 import { Text } from "@/shared/ui/text";
-
+import { useDiaryEmotionStore } from "@/store/diary-emotion/useDiaryEmotionStore";
 import { EMOTIONS } from "../model/emotion";
-import { useDiaryEmotionStore } from "../model/useDiaryEmotionStore";
 import { TagList } from "./TagList";
 
 interface SituationStepProps {
@@ -23,8 +23,9 @@ const getEmotionValue = (emotionId: string | null): number => {
 export const SituationStep = ({ onValidChange }: SituationStepProps): React.ReactElement => {
   const { selectedEmotionId, selectedSituationIds, setSelectedSituationIds } =
     useDiaryEmotionStore();
-  const { data } = useEmotionTagsQuery(getEmotionValue(selectedEmotionId));
-  const tags = data?.tags.map((tag) => ({ id: String(tag.id), label: tag.label })) ?? [];
+  const { data, isError } = useEmotionTagsQuery(getEmotionValue(selectedEmotionId));
+  const sourceTags = isError ? MOCK_SITUATIONS : (data?.tags ?? []);
+  const tags = sourceTags.map((tag) => ({ id: String(tag.id), label: tag.label }));
 
   useEffect(() => {
     onValidChange(selectedSituationIds.length === 0);
